Fix playlistApp name to playlist
For actions and reducers only.
