+++ src/jsx/components/playlistApp/player/Player.jsx
@@ -9,20 +9,20 @@ import UserWidget from 'components/generics/UserWidget'
 import ManageButton from './ManageButton'
 import PlayerNotification from './Notification'
 import { IsPlaylistManagerOrOwner } from 'components/permissions/Playlist'
-import { sendPlayerCommand } from 'actions/playlistApp'
-import { playlistAppDigestPropType, playerCommandsPropType } from 'reducers/playlistApp'
+import { sendPlayerCommand } from 'actions/playlist'
+import { playlistDigestPropType, playerCommandsPropType } from 'reducers/playlist'
 import { alterationStatusPropType } from 'reducers/alterationsStatus'
 
 class Player extends Component {
     static propTypes = {
-        playlistAppDigest: playlistAppDigestPropType.isRequired,
+        playlistDigest: playlistDigestPropType.isRequired,
         sendPlayerCommandsStatus: PropTypes.objectOf(alterationStatusPropType),
         sendPlayerCommand: PropTypes.func.isRequired,
     }
 
     render() {
-        const { player_status, player_manage, player_errors } = this.props.playlistAppDigest.data
-        const { fetchError } = this.props.playlistAppDigest
+        const { player_status, player_manage, player_errors } = this.props.playlistDigest.data
+        const { fetchError } = this.props.playlistDigest
         const isPlaying = !!player_status.playlist_entry
         const controlDisabled = !isPlaying || fetchError
 
@@ -148,7 +148,7 @@ class Player extends Component {
 }
 
 const mapStateToProps = (state) => ({
-    playlistAppDigest: state.playlistApp.digest,
+    playlistDigest: state.playlist.digest,
     sendPlayerCommandsStatus: state.alterationsStatus.sendPlayerCommands,
 })
