+++ src/jsx/reducers/playlist.js
@@ -1,18 +1,118 @@
+import { combineReducers } from 'redux'
 import PropTypes from 'prop-types'
-import { PLAYLIST_REQUEST, PLAYLIST_SUCCESS, PLAYLIST_FAILURE } from 'actions/playlistApp'
-import { PLAYLIST_PLAYED_REQUEST, PLAYLIST_PLAYED_SUCCESS, PLAYLIST_PLAYED_FAILURE } from 'actions/playlistApp'
-import { PLAYLIST_PLAYED_ADD, PLAYLISTAPP_DIGEST_SUCCESS } from 'actions/playlistApp'
-import { playlistEntryPropType, playlistPlayedEntryPropType } from 'serverPropTypes/playlist'
+import { PLAYLISTAPP_DIGEST_REQUEST, PLAYLISTAPP_DIGEST_SUCCESS, PLAYLISTAPP_DIGEST_FAILURE } from 'actions/playlist'
+import { PLAYLIST_REQUEST, PLAYLIST_SUCCESS, PLAYLIST_FAILURE } from 'actions/playlist'
+import { PLAYLIST_PLAYED_REQUEST, PLAYLIST_PLAYED_SUCCESS, PLAYLIST_PLAYED_FAILURE } from 'actions/playlist'
+import { PLAYLIST_PLAYED_ADD } from 'actions/playlist'
+import { FORM_SUCCESS } from 'actions/forms'
+import { ALTERATION_SUCCESS } from 'actions/alterationsStatus'
 import { Status } from './alterationsStatus'
+import { playerStatusPropType, playerManagePropType, playerErrorPropType, karaStatusPropType } from 'serverPropTypes/playlist'
+import { playlistEntryPropType, playlistPlayedEntryPropType } from 'serverPropTypes/playlist'
 import { updateData } from 'utils'
+
 
 /**
  * This reducer contains playlist related state
  */
 
+
 /**
- * Playlist from server
+ * Player information digest from server
  */
+
+
+export const playlistDigestPropType = PropTypes.shape({
+    status: PropTypes.symbol,
+    data: PropTypes.shape({
+        player_status: playerStatusPropType.isRequired,
+        player_manage: playerManagePropType.isRequired,
+        player_errors: PropTypes.arrayOf(playerErrorPropType).isRequired,
+        kara_status: karaStatusPropType.isRequired,
+    }).isRequired,
+})
+
+const defaultPlaylistAppDigest = {
+    status: null,
+    data: {
+        player_status: {
+            playlist_entry: null,
+            timing: 0
+        },
+        player_manage: {
+            pause: false,
+            skip: false
+        },
+        player_errors: [],
+        kara_status: {
+            status: null,
+        },
+    },
+}
+
+function digest(state = defaultPlaylistAppDigest, action) {
+    switch (action.type) {
+        case PLAYLISTAPP_DIGEST_REQUEST:
+            return {
+                ...state,
+                status: state.status || Status.pending
+            }
+
+        case PLAYLISTAPP_DIGEST_SUCCESS:
+            return {
+                status: Status.successful,
+                data: action.response,
+            }
+
+        case PLAYLISTAPP_DIGEST_FAILURE:
+            return {
+                ...state,
+                status: Status.failed,
+            }
+
+        // if a pause command has been successfuly sent to the server,
+        // adapt the state now
+        case ALTERATION_SUCCESS:
+            if (action.alterationName === 'sendPlayerCommand' &&
+                action.elementId === 'pause') {
+                return {
+                    ...state,
+                    data: {
+                        ...state.data,
+                        player_manage: {
+                            ...state.data.player_manage,
+                            pause: action.value,
+                        }
+                    }
+                }
+            }
+
+            return state
+
+        // if the kara status has been successfuly edited, adapt the state now
+        case FORM_SUCCESS:
+            if (action.formName == "editKaraStatus") {
+                return {
+                    ...state,
+                    data: {
+                        ...state.data,
+                        kara_status: {
+                            status: action.response.status
+                        }
+                    }
+                }
+            }
+
+        default:
+            return state
+    }
+}
+
+
+/**
+ * Playlist of entries from server
+ */
+
 
 export const playlistEntriesStatePropType = PropTypes.shape({
     status: PropTypes.symbol,
@@ -30,7 +130,7 @@ const defaultEntries = {
     },
 }
 
-export function entries(state = defaultEntries, action) {
+function entries(state = defaultEntries, action) {
     switch (action.type) {
         case PLAYLIST_REQUEST:
             return {
@@ -55,6 +155,12 @@ export function entries(state = defaultEntries, action) {
     }
 }
 
+
+/**
+ * Playlist of played entries from server
+ */
+
+
 export const playlistPlayedEntriesStatePropType = PropTypes.shape({
     status: PropTypes.symbol,
     data: PropTypes.shape({
@@ -73,7 +179,7 @@ const defaultPlayedEntries = {
     },
 }
 
-export function playedEntries(state = defaultPlayedEntries, action) {
+function playedEntries(state = defaultPlayedEntries, action) {
     switch (action.type) {
         case PLAYLIST_PLAYED_REQUEST:
             return {
@@ -119,3 +225,15 @@ export function playedEntries(state = defaultPlayedEntries, action) {
             return state
     }
 }
+
+/**
+ * Playlist
+ */
+
+const playlist = combineReducers({
+    digest,
+    entries,
+    playedEntries,
+})
+
+export default playlist
